Revert to using FirstOrDefault
Using single didn't work because in the case that the DC=null and
it runs to the second part of the statement there can be situations
where more than one project exsits that uses that health risk code.
In this event this statement will set the healthriskId to the First
one that matches, but when you edit a report with unknown sender
you get a chance to rectify it when you set a DC. Then the report
gets the health risk code that corresponds to the DCs project.